pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta" 

[project]
name = "tvd_coverage"
version = "1.0"
dependencies = [
  "pip >= 23.2.1",
  "numpy >= 1.26.0",
  "scipy >= 1.11.1",
  "matplotlib >= 3.7.3",
  "pyvoro-mmalahe >= 1.3.4",
  "tqdm >= 4.65.0"
]
requires-python = ">= 3.11"
authors = [
  {name = "Brandon Bao", email = "bjbao@ucsd.edu"},
]
description = "Singular Perturbation Coverage Control Paper Code."
readme = "README.md"
license = {file = "LICENSE"}

[project.urls]
Repository = "https://github.com/me/spam.git"

[tool.setuptools]
packages = ["pycoverage_robotarium", "pycoverage3d"]